refactor: changed how the background p5js sketch is rendered
By loading the background sketch as I was before, via the p5js cdn library and a sketch script in
the page, sometimes the sketch wouldn't load due to sveltekit's pre-rendering and server side
rendering features. Instead, I opted to install the p5js library as a dependency and use  p5-svelte
found here: https://github.com/tonyketcham/p5-svelte

BREAKING CHANGE: p5 and p5-svelte must now be installed as dependencies in order to build the site.
