+++ package.json
@@ -15,6 +15,8 @@
 		"autoprefixer": "^10.4.13",
 		"commitizen": "^4.3.0",
 		"cz-conventional-changelog": "^3.3.0",
+		"p5": "^1.5.0",
+		"p5-svelte": "^3.1.2",
 		"postcss": "^8.4.21",
 		"svelte": "^3.54.0",
 		"tailwindcss": "^3.2.4",
