+++ src/routes/+page.svelte
@@ -1,6 +1,11 @@
 <script>
+  import BackgroundComponent from '../lib/components/BackgroundComponent.svelte';
 </script>
 
 <svelte:head>
   <title>Hello, I'm Chris Sterza</title>
 </svelte:head>
+
+<main>
+  <BackgroundComponent />
+</main>
